+++ src/components/eholdings/eholdings.js
@@ -31,7 +31,7 @@ import NoteEdit from '../../routes/note-edit';
 
 import SettingsCustomLabelsRoute from '../../routes/settings-custom-labels';
 import SettingsKnowledgeBaseRoute from '../../routes/settings-knowledge-base';
-import SettingsRootProxyRoute from '../../routes/settings-root-proxy';
+import SettingsRootProxyRoute from '../../routes/settings-root-proxy-route';
 import SettingsAccessStatusTypesRoute from '../../routes/settings-access-status-types';
 import SettingsAssignedUsersRoute from '../../routes/settings-assigned-users-route';
 import SettingsUsageConsolidationRoute from '../../routes/settings-usage-consolidation-route';