+++ src/components/eholdings/eholdings.js
@@ -29,7 +29,7 @@ import NoteCreate from '../../routes/note-create';
 import NoteView from '../../routes/note-view';
 import NoteEdit from '../../routes/note-edit';
 
-import SettingsCustomLabelsRoute from '../../routes/settings-custom-labels';
+import SettingsCustomLabelsRoute from '../../routes/custom-labels-route';
 import SettingsKnowledgeBaseRoute from '../../routes/settings-knowledge-base';
 import SettingsRootProxyRoute from '../../routes/settings-root-proxy';
 import SettingsAccessStatusTypesRoute from '../../routes/settings-access-status-types-route';