+++ src/components/eholdings/eholdings.js
@@ -32,7 +32,7 @@ import NoteEdit from '../../routes/note-edit';
 import SettingsCustomLabelsRoute from '../../routes/custom-labels-route';
 import SettingsKnowledgeBaseRoute from '../../routes/settings-knowledge-base';
 import SettingsRootProxyRoute from '../../routes/settings-root-proxy';
-import SettingsAccessStatusTypesRoute from '../../routes/settings-access-status-types';
+import SettingsAccessStatusTypesRoute from '../../routes/settings-access-status-types-route';
 import SettingsAssignedUsersRoute from '../../routes/settings-assigned-users-route';
 import SettingsUsageConsolidationRoute from '../../routes/settings-usage-consolidation-route';