+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 * Add tests for SearchRoute component. (UIEH-1192)
 * Shift focus once user hits Previous or Next pagination link. (UIEH-1206)
 * Allow user to save special characters in Settings Custom labels. (UIEH-1209)
+* Remove SearchPane component. (UIEH-1177)
 
 ## [6.1.1] (https://github.com/folio-org/ui-eholdings/tree/v6.1.1) (2021-07-15)
 * Fix: saving the update of coverage settings temporarily shows the old date. (UIEH-1146)