+++ CHANGELOG.md
@@ -70,6 +70,7 @@
 * Allow user to save special characters in Settings Custom labels. (UIEH-1209)
 * Remove SearchPane component. (UIEH-1177)
 * Combine QueryList and QuerySearchList components into one. (UIEH-1208)
+* Jest+RTL > Test SettingsRootProxyRoute. (UIEH-1197)
 
 ## [6.1.1] (https://github.com/folio-org/ui-eholdings/tree/v6.1.1) (2021-07-15)
 * Fix: saving the update of coverage settings temporarily shows the old date. (UIEH-1146)